Fixed EBC build issues. Fixed ICC build issues.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@3431 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Library/DxeNetLib/Netbuffer.c
@@ -1711,7 +1711,7 @@ NetbufChecksum (
       // The checksum starts with an odd byte, swap
       // the checksum before added to total checksum
       //
-      BlockSum = NET_SWAP_SHORT (BlockSum);
+      BlockSum = (UINT16) NET_SWAP_SHORT (BlockSum);
     }
 
     TotalSum = NetAddChecksum (BlockSum, TotalSum);
